feat: SQLExecuteQueryOperator output processors renamed and moved
